refactor(training): rebuild Training page from scratch
Complete rewrite of src/app/activities/page.tsx (~900 lines, was 2250).
Old file saved as page.backup.tsx.

New structure:
- Left sidebar on desktop (≥768px) / dropdown menu on mobile
- Three sections: Données | Analyse | Progression
- Clean, minimal design — dark theme, inline styles, no old logic

Données: stat cards (30d), weekly TSS bar chart (12w), sport breakdown
Analyse: searchable + filterable activity list, detail panel with altitude sparkline + laps
Progression: personal records per sport, recent 90-day highlights
